Adds  to the input list in Functions/PrepareDatauploading.R
Now Source is an input rather than to be inferred by the code. Practically in this format
only data of a one source can be handelled. So it is recommended to use one data file at a time.

In the previous version Source used to be ineferred, hence, data files with different Sources could be
handelled in a single run of the function.

This change has been made to make the structure of the EarnixUploader more compatible with the PricingPlatforms.